Frontend/requirements.txt
requests
Flask
Flask-WTF
aws-xray-sdk
flask-login
boto3
python_jose
Flask-Uploads